优化项目结构，对navbar 进行解构
