+++ .env.development
@@ -10,11 +10,3 @@ VITE_PROXY_DOMAIN = /api
 # 开发环境路由历史模式
 VITE_ROUTER_HISTORY = "hash"
 
-# 开发环境后端地址
-VITE_ASSIS_INTERFACE_ADDRESS = "https://assis-serv.linkcloud-energy.com/"
-VITE_PROD_INTERFACE_ADDRESS = "https://vpp.energyfocus.cloud/"
-
-# mqtt地址
-VITE_MQTT_ADDRESS = "assis-serv.linkcloud-energy.com"
-
-VITE_MQTT_CLIENT_ID = "web_"
