+++ .env.production
@@ -6,12 +6,3 @@ VITE_BASE_URL = ./
 
 # 线上环境路由历史模式
 VITE_ROUTER_HISTORY = "hash"
-
-# 线上环境后端地址
-VITE_ASSIS_INTERFACE_ADDRESS = "https://vpp.energyfocus.cloud/"
-VITE_PROD_INTERFACE_ADDRESS = ""
-
-# mqtt地址
-VITE_MQTT_ADDRESS = "vpp.energyfocus.cloud"
-
-VITE_MQTT_CLIENT_ID = "web_"
