+++ src/layouts/DefaultLayout.vue
@@ -10,5 +10,5 @@
 </template>
 
 <script setup>
-import NavBar from './components/NavBar.vue';
+import NavBar from './components/navbar/NavBar.vue';
 </script>
